E-COMMERCE-STORE/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.7.2
python-dotenv==1.0.0
razorpay==1.3.0
pydantic==2.9.2
starlette==0.27.0
PyJWT==2.8.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
aiofiles==23.2.1
cloudinary==1.41.0